Implement app settings API
