+++ src/commands/appSettings/downloadAppSettings.ts
@@ -12,6 +12,7 @@ import { localSettingsFileName, viewOutput } from "../../constants";
 import { ext } from "../../extensionVariables";
 import { getLocalSettingsJson, ILocalSettingsJson } from "../../funcConfig/local.settings";
 import { localize } from "../../localize";
+import { ISimpleAppSettingsClient } from "../../vscode-azurefunctions.api";
 import { confirmOverwriteSettings } from "./confirmOverwriteSettings";
 import { decryptLocalSettings } from "./decryptLocalSettings";
 import { encryptLocalSettings } from "./encryptLocalSettings";
@@ -23,7 +24,12 @@ export async function downloadAppSettings(context: IActionContext, node?: AppSet
     }
 
     const client: IAppSettingsClient = node.client;
+    await node.runWithTemporaryDescription(localize('downloading', 'Downloading...'), async () => {
+        await downloadAppSettingsInternal(context, client);
+    });
+}
 
+export async function downloadAppSettingsInternal(context: IActionContext, client: ISimpleAppSettingsClient): Promise<void> {
     const message: string = localize('selectLocalSettings', 'Select the destination file for your downloaded settings.');
     const localSettingsPath: string = await getLocalSettingsFile(message);
     const localSettingsUri: vscode.Uri = vscode.Uri.file(localSettingsPath);
@@ -48,10 +54,9 @@ export async function downloadAppSettings(context: IActionContext, node?: AppSet
             await confirmOverwriteSettings(remoteSettings.properties, localSettings.Values, localSettingsFileName);
         }
 
-        await node.runWithTemporaryDescription(localize('downloading', 'Downloading...'), async () => {
-            await fse.ensureFile(localSettingsPath);
-            await fse.writeJson(localSettingsPath, localSettings, { spaces: 2 });
-        });
+        await fse.ensureFile(localSettingsPath);
+        await fse.writeJson(localSettingsPath, localSettings, { spaces: 2 });
+
     } finally {
         if (isEncrypted) {
             await encryptLocalSettings(context, localSettingsUri);
